+++ src/Simplify.Mail/IMailSender.cs
@@ -1,6 +1,9 @@
-﻿using System.Collections.Generic;
-using System.Net.Mail;
+using System;
+using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
+using MailKit.Net.Smtp;
+using MimeKit;
 using Simplify.Mail.Settings;
 
 namespace Simplify.Mail;
@@ -8,7 +11,7 @@ namespace Simplify.Mail;
 /// <summary>
 /// Represent E-mail sending interface.
 /// </summary>
-public interface IMailSender
+public interface IMailSender : IDisposable
 {
 	/// <summary>
 	/// MailSender settings.
@@ -25,31 +28,33 @@ public interface IMailSender
 	/// Send single e-mail.
 	/// </summary>
 	/// <param name="client">Smtp client.</param>
-	/// <param name="mailMessage">The mail message.</param>
+	/// <param name="mimeMessage">The MIME message.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
-	void Send(SmtpClient client, MailMessage mailMessage, string bodyForAntiSpam = null);
+	void Send(SmtpClient client, MimeMessage mimeMessage, string bodyForAntiSpam = null);
 
 	/// <summary>
 	/// Send single e-mail asynchronously.
 	/// </summary>
 	/// <param name="client">Smtp client.</param>
-	/// <param name="mailMessage">The mail message.</param>
+	/// <param name="mimeMessage">The MIME message.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
-	Task SendAsync(SmtpClient client, MailMessage mailMessage, string bodyForAntiSpam = null);
+	/// <param name="cancellationToken">The cancellation token.</param>
+	Task SendAsync(SmtpClient client, MimeMessage mimeMessage, string bodyForAntiSpam = null, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// Send single e-mail.
 	/// </summary>
-	/// <param name="mailMessage">The mail message.</param>
+	/// <param name="mimeMessage">The MIME message.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
-	void Send(MailMessage mailMessage, string bodyForAntiSpam = null);
+	void Send(MimeMessage mimeMessage, string bodyForAntiSpam = null);
 
 	/// <summary>
 	/// Send single e-mail asynchronously.
 	/// </summary>
-	/// <param name="mailMessage">The mail message.</param>
+	/// <param name="mimeMessage">The MIME message.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
-	Task SendAsync(MailMessage mailMessage, string bodyForAntiSpam = null);
+	/// <param name="cancellationToken">The cancellation token.</param>
+	Task SendAsync(MimeMessage mimeMessage, string bodyForAntiSpam = null, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// Send single e-mail.
@@ -65,7 +70,7 @@ public interface IMailSender
 	/// Process status, <see langword="true" /> if message is processed to sent successfully
 	/// </returns>
 	void Send(SmtpClient client, string from, string to, string subject, string body, string bodyForAntiSpam = null,
-		params Attachment[] attachments);
+		params MimeEntity[] attachments);
 
 	/// <summary>
 	/// Send single e-mail asynchronously.
@@ -77,11 +82,12 @@ public interface IMailSender
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
+	/// <param name="cancellationToken">The cancellation token.</param>
 	/// <returns>
 	/// Process status, <see langword="true" /> if message is processed to sent successfully
 	/// </returns>
 	Task SendAsync(SmtpClient client, string from, string to, string subject, string body, string bodyForAntiSpam = null,
-		params Attachment[] attachments);
+		MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// Send single e-mail.
@@ -94,7 +100,7 @@ public interface IMailSender
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <returns>Process status, <see langword="true"/> if message is processed to sent successfully</returns>
 	void Send(string from, string to, string subject, string body, string bodyForAntiSpam = null,
-		params Attachment[] attachments);
+		params MimeEntity[] attachments);
 
 	/// <summary>
 	/// Send single e-mail asynchronously.
@@ -105,9 +111,10 @@ public interface IMailSender
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
+	/// <param name="cancellationToken">The cancellation token.</param>
 	/// <returns>Process status, <see langword="true"/> if message is processed to sent successfully</returns>
 	Task SendAsync(string from, string to, string subject, string body, string bodyForAntiSpam = null,
-		params Attachment[] attachments);
+		MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients in one e-mail.
@@ -121,7 +128,7 @@ public interface IMailSender
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	void Send(SmtpClient client, string fromMailAddress, IList<string> addresses, string subject, string body,
-		string bodyForAntiSpam = null, params Attachment[] attachments);
+		string bodyForAntiSpam = null, params MimeEntity[] attachments);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients in one e-mail asynchronously.
@@ -133,9 +140,10 @@ public interface IMailSender
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
+	/// <param name="cancellationToken">The cancellation token.</param>
 	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	Task SendAsync(SmtpClient client, string fromMailAddress, IList<string> addresses, string subject, string body,
-		string bodyForAntiSpam = null, params Attachment[] attachments);
+		string bodyForAntiSpam = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients in one e-mail.
@@ -148,7 +156,7 @@ public interface IMailSender
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	void Send(string fromMailAddress, IList<string> addresses, string subject, string body,
-		string bodyForAntiSpam = null, params Attachment[] attachments);
+		string bodyForAntiSpam = null, params MimeEntity[] attachments);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients in one e-mail asynchronously.
@@ -159,9 +167,10 @@ public interface IMailSender
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
+	/// <param name="cancellationToken">The cancellation token.</param>
 	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	Task SendAsync(string fromMailAddress, IList<string> addresses, string subject, string body,
-		string bodyForAntiSpam = null, params Attachment[] attachments);
+		string bodyForAntiSpam = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients and carbon copy recipients in one e-mail.
@@ -176,7 +185,7 @@ public interface IMailSender
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	void Send(SmtpClient client, string fromMailAddress, IList<string> addresses, IList<string> ccAddresses,
-		string subject, string body, string bodyForAntiSpam = null, params Attachment[] attachments);
+		string subject, string body, string bodyForAntiSpam = null, params MimeEntity[] attachments);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients and carbon copy recipients in one e-mail asynchronously.
@@ -189,9 +198,10 @@ public interface IMailSender
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
+	/// <param name="cancellationToken">The cancellation token.</param>
 	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	Task SendAsync(SmtpClient client, string fromMailAddress, IList<string> addresses, IList<string> ccAddresses,
-		string subject, string body, string bodyForAntiSpam = null, params Attachment[] attachments);
+		string subject, string body, string bodyForAntiSpam = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients and carbon copy recipients in one e-mail.
@@ -204,7 +214,8 @@ public interface IMailSender
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
-	void Send(string fromMailAddress, IList<string> addresses, IList<string> ccAddresses, string subject, string body, string bodyForAntiSpam = null, params Attachment[] attachments);
+	void Send(string fromMailAddress, IList<string> addresses, IList<string> ccAddresses, string subject, string body,
+		string bodyForAntiSpam = null, params MimeEntity[] attachments);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients and carbon copy recipients in one e-mail asynchronously.
@@ -216,9 +227,10 @@ public interface IMailSender
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
+	/// <param name="cancellationToken">The cancellation token.</param>
 	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	Task SendAsync(string fromMailAddress, IList<string> addresses, IList<string> ccAddresses, string subject, string body,
-		string bodyForAntiSpam = null, params Attachment[] attachments);
+		string bodyForAntiSpam = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients separately
@@ -232,7 +244,7 @@ public interface IMailSender
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	void SendSeparately(SmtpClient client, string fromMailAddress, IList<string> addresses, string subject, string body,
-		string bodyForAntiSpam = null, params Attachment[] attachments);
+		string bodyForAntiSpam = null, params MimeEntity[] attachments);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients separately
@@ -244,9 +256,10 @@ public interface IMailSender
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
+	/// <param name="cancellationToken">The cancellation token.</param>
 	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	Task SendSeparatelyAsync(SmtpClient client, string fromMailAddress, IList<string> addresses, string subject, string body,
-		string bodyForAntiSpam = null, params Attachment[] attachments);
+		string bodyForAntiSpam = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients separately
@@ -259,7 +272,7 @@ public interface IMailSender
 	/// <param name="attachments">The attachments to an e-mail.</param>
 	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	void SendSeparately(string fromMailAddress, IList<string> addresses, string subject, string body,
-		string bodyForAntiSpam = null, params Attachment[] attachments);
+		string bodyForAntiSpam = null, params MimeEntity[] attachments);
 
 	/// <summary>
 	/// Send e-mail to multiple recipients separately
@@ -270,7 +283,8 @@ public interface IMailSender
 	/// <param name="body">e-mail body.</param>
 	/// <param name="bodyForAntiSpam">Part of an e-mail body just for anti-spam checking.</param>
 	/// <param name="attachments">The attachments to an e-mail.</param>
+	/// <param name="cancellationToken">The cancellation token.</param>
 	/// <returns>Process status, <see langword="true"/> if all messages are processed to sent successfully</returns>
 	Task SendSeparatelyAsync(string fromMailAddress, IList<string> addresses, string subject, string body,
-		string bodyForAntiSpam = null, params Attachment[] attachments);
+		string bodyForAntiSpam = null, MimeEntity[] attachments = null, CancellationToken cancellationToken = default);
 }